+++ frontend/src/App.tsx
@@ -1,16 +1,18 @@
 import "./App.css";
 
-import { useReduce } from "./Reducer-context";
+import { useReduce } from "./reducer-context/Reducer-context";
 
-import { Question } from "./Question";
-import { Routes, Route } from "react-router-dom";
-import { Login } from "./Login";
-import { Home } from "./Home";
-import { Privateroute } from "./Privateroute";
-import { Learninggraph } from "./Learninggraph";
-import { Navbar } from "./Navbar";
-import { User } from "./User";
-import {Signup} from "./Signup";
+// import { Question } from "./pages/Question";
+ import { Routes, Route } from "react-router-dom";
+// import { Login } from "./pages/Login";
+// import { Home } from "./pages/Home";
+ import { Privateroute } from "./Privateroute";
+// import { Learninggraph } from "./pages/Learninggraph";
+ //import { Navbar } from "./components/Navbar";
+// import { User } from "./pages/User";
+// import {Signup} from "./pages/Signup";
+import {Question,Login,LearningGraph,Home,User,Signup } from "./pages"
+import {Navbar} from "./components"
 export default function App() {
   console.log("render");
   console.log(useReduce());
@@ -19,7 +21,7 @@ export default function App() {
       <Navbar />
       <Routes>
         <Route path="/" element={<Home />} />
-        <Privateroute path="/learningcurve/:category" element={<Learninggraph />} />
+        <Privateroute path="/learningcurve/:category" element={<LearningGraph />} />
         <Route path="/login" element={<Login />} />
         <Route path="/signup" element={<Signup />} />
         <Privateroute path="/user" element={<User />} />
