+++ frontend/src/components/index.ts
@@ -1,0 +1,6 @@
+export {Answer} from "./Answer"
+export {CircularProgressWithLabel} from "./CircularProgressWithLabel"
+export {Navbar} from "./Navbar"
+export {Qsnblock} from "./Qsnblock"
+export {Statistics} from "./Statistics"
+export {Header} from "./Header"
