+++ frontend/src/pages/index.tsx
@@ -1,0 +1,6 @@
+export {Home} from "./Home"
+export {LearningGraph} from "./LearningGraph"
+export {Login} from "./Login"
+export {Question} from "./Question"
+export {Signup} from "./Signup"
+export {User} from "./User"
